requirements.txt
selenium==4.5.0
baidu_api
baidu_aip
Pillow
threadpool
chardet
yaml